Shorten EuclideanMST module name.
