+++ hgeometry/hgeometry.cabal
@@ -134,6 +134,7 @@ library
 
                     Algorithms.Geometry.PolyLineSimplification.DouglasPeucker
 
+                    Algorithms.Geometry.EuclideanMST
                     Algorithms.Geometry.EuclideanMST.EuclideanMST
 
                     Algorithms.Geometry.WellSeparatedPairDecomposition.WSPD
